+++ package.json
@@ -20,6 +20,7 @@
       "react-dom": "^18.2.0",
       "react-router-dom": "^6.3.0",
       "react-scripts": "5.0.1",
+      "swr": "^1.3.0",
       "typescript": "^4.7.4",
       "web-vitals": "^2.1.0"
    },
